Update README.md
Joseph Marie Charles dit (called or nicknamed) Jacquard (French: [ʒakaʁ]; 7 July 1752 – 7 August 1834) was a French weaver and merchant. He played an important role in the development of the earliest programmable loom (the "Jacquard loom"), which in turn played an important role in the development of other programmable machines, such as an early version of digital compiler used by IBM to develop the modern day computer.